fix(lenses): Update defaultSource value when lens params change
ensureDefaultSource only creates sources if they don't exist, but doesn't
update the value if they do. This caused lens parameter edits to be ignored
since the source already existed from when the lens was added.

Now we call setDefaultValue after ensureDefaultSource to ensure the value
is always updated when converting LensDefinition to LensInstance.